+++ src/main/java/net/frozenblock/lib/entity/mixin/command/WardenSpawnTrackerCommandMixin.java
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-package net.frozenblock.lib.core.mixin;
+package net.frozenblock.lib.entity.mixin.command;
 
 import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;