+++ CHANGELOG.md
@@ -11,3 +11,4 @@ Put changelog here:
 - Fixed Apparition `RenderType`s, and cleaned up existing ones.
 - Added `ShaderRegistryAPI`, allowing modders to register custom shaders.
   - This API is not present past 1.21.1 as Vanilla gained this functionality in 1.21.2.
+- Added `ItemTooltipAdditionAPI`, letting modders append custom text tooltips to items when certain conditions are met.